Sensor Project/code/requirements.txt
numpy>=1.20.0
pandas>=1.3.0
paho-mqtt>=1.6.1
influxdb-client>=1.24.0
pyserial>=3.5
flask>=2.0.1
RPi.GPIO>=0.7.0
spidev>=3.5
matplotlib>=3.4.0
scipy>=1.7.0
pyyaml>=6.0
